// enums.ts
export enum LoanStatus {
	Current = "CURRENT",
	Paid = "PAID",
	PaidLate = "PAID_LATE",
}

export enum LocationTrait {
	AbundantMetalOres = "ABUNDANT_METAL_ORES",
	AbundantNaturalChemicals = "ABUNDANT_NATURAL_CHEMICALS",
	ArableLand = "ARABLE_LAND",
	Helium3 = "HELIUM_3",
	MetalOres = "METAL_ORES",
	NaturalChemicals = "NATURAL_CHEMICALS",
	SomeArableLand = "SOME_ARABLE_LAND",
	SomeHelium3 = "SOME_HELIUM_3",
	SomeMetalOres = "SOME_METAL_ORES",
	SomeNaturalChemicals = "SOME_NATURAL_CHEMICALS",
	SomeRareMetalOres = "SOME_RARE_METAL_ORES",
	SomeTechnologicalRuins = "SOME_TECHNOLOGICAL_RUINS",
	RareMetalOres = "RARE_METAL_ORES",
	TechnologicalRuins = "TECHNOLOGICAL_RUINS",
}

export enum LocationType {
	Anomaly = "ANOMALY",
	Asteroid = "ASTEROID",
	GasGiant = "GAS_GIANT",
	Moon = "MOON",
	Nebula = "NEBULA",
	Planet = "PLANET",
	Wormhole = "WORMHOLE",
}

export enum StructureType {
	ChemicalPlant = "CHEMICAL_PLANT",
	DroneFactory = "DRONE_FACTORY",
	ElectronicsFactory = "ELECTRONICS_FACTORY",
	ExplosivesFacility = "EXPLOSIVES_FACILITY",
	FabricationPlant = "FABRICATION_PLANT",
	Farm = "FARM",
	FuelRefinery = "FUEL_REFINERY",
	Mine = "MINE",
	RareEarthMine = "RARE_EARTH_MINE",
	ResearchOutpost = "RESEARCH_OUTPOST",
	Shipyard = "SHIPYARD",
}

export enum HTTPMethod {
    GET = "GET",
    POST = "POST",
    PUT = "PUT",
    DELETE = "DELETE"
}